requirements.txt
# Action Grounding Research - Dependencies
# =========================================
# Install with: pip install -r requirements.txt

# =============================================================================
# Core ML (PyTorch/Transformers - GPU compatible)
# =============================================================================
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.28.0
bitsandbytes>=0.43.0  # For 4-bit and 8-bit quantization

# =============================================================================
# Configuration & Data Validation
# =============================================================================
pydantic>=2.6.0
pydantic-settings>=2.2.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# =============================================================================
# Data Processing & Analysis
# =============================================================================
numpy>=1.26.0
pandas>=2.2.0
pyarrow>=15.0.0  # For Parquet support
scikit-learn>=1.4.0

# =============================================================================
# Visualization
# =============================================================================
matplotlib>=3.8.0
seaborn>=0.13.0

# =============================================================================
# Notebooks & Progress
# =============================================================================
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0
tqdm>=4.66.0

# =============================================================================
# API Clients
# =============================================================================
openai>=1.12.0
httpx>=0.27.0

# =============================================================================
# Optional: Apple Silicon (MLX) - Uncomment if on Mac
# =============================================================================
# mlx>=0.21.0
# mlx-lm>=0.19.0

# =============================================================================
# Optional: vLLM for faster inference (2-3x faster than PyTorch)
# Install separately if needed: pip install vllm
# Note: vLLM has large dependencies (~5GB), install only if using vLLM backend
# =============================================================================
# vllm>=0.4.0  # Uncomment to install vLLM (requires significant disk space)

# =============================================================================
# Development & Testing
# =============================================================================
# pytest>=8.0.0
# pytest-cov>=4.1.0
# black>=24.2.0
# ruff>=0.3.0
